Bug 1377607 Cast int to unsigned to silence an error about signed vs unsigned comparisons r=froydnj
MozReview-Commit-ID: Op3tiipqcr

UltraBlame original commit: c2c9baffd57b3d8d200b3f34fc88332cf9b68ee4

+++ mozglue/misc/TimeStamp_windows.cpp
@@ -478,7 +478,7 @@ HasStableTSC()
 
   
   __cpuid(regs, 0x80000000);
-  if (regs[0] < 0x80000007) {
+  if ((unsigned int)regs[0] < 0x80000007) {
     
     
     return false;
